refactor to make file open opts as int instead of long

+++ core/src/main/c/share/files.c
@@ -176,13 +176,13 @@ JNIEXPORT jint JNICALL Java_io_questdb_std_Files_openRW
 }
 
 JNIEXPORT jint JNICALL Java_io_questdb_std_Files_openRWOpts
-        (JNIEnv *e, jclass cl, jlong lpszName, jlong opts) {
+        (JNIEnv *e, jclass cl, jlong lpszName, jint opts) {
     umask(0);
     return open((const char *) lpszName, O_CREAT | O_RDWR | opts, 0644);
 }
 
 JNIEXPORT jint JNICALL Java_io_questdb_std_Files_openRWOptsNoCreate
-        (JNIEnv *e, jclass cl, jlong lpszName, jlong opts) {
+        (JNIEnv *e, jclass cl, jlong lpszName, jint opts) {
     umask(0);
     return open((const char *) lpszName, O_RDWR | opts, 0644);
 }
